+++ paper/paper.bib
@@ -124,7 +124,7 @@ number = "3",
 pages = "136 - 141",
 year = "1996",
 issn = "0263-7855",
-doi = "https://doi.org/10.1016/S0263-7855(96)00043-4",
+doi = "10.1016/S0263-7855(96)00043-4",
 url = "http://www.sciencedirect.com/science/article/pii/S0263785596000434",
 author = "W. Smith and T.R. Forester",
 abstract = "DL_POLY_2.0 is a general-purpose parallel molecular dynamics simulation package developed at Daresbury Laboratory under the auspices of the Council for the Central Laboratory of the Research Councils. Written to support academic research, it has a wide range of applications and is designed to run on a wide range of computers: from single processor workstations to parallel supercomputers. Its structure, functionality, performance, and availability are described."
