### 2020-10-01 #2709 @xuanxu
> openjournals/dev can you see why this paper is not compiling?

whedon is having problems lately when compiling the pdf at the same time it is doing other checks. Retriyng when other task are finished should work.
