### 2020-10-01 8481fd4
Update paper.md
