### 2020-10-01 #2709 @symmy596
I have noticed that the figure has not rendered. I have updated the markdown.
