### 2020-10-01 #2709 @Kevin-Mattheus-Moerman
@symmy596 thanks for your submission. I've flagged this submission for additional review by our board, to check if it is in scope for JOSS. In particular it looks like it is close to being too minor (to me at first sight) in terms of functionality / the amount of scholarly work it contains. This internal review will take about a week. In the mean time if you could clearly and concisely outline the different types of functionality this work offers and how these are significant or novel contributions, and how they may differ in relation to other work, that may help us judge the scope of this work. Thanks.
