### 2020-10-01 #2709 @symmy596
polypy was the analysis software used in three pieces of published work during my PhD (10.1088/2515-7655/ab28b5, 10.1098/rsta.2019.0026 and 10.1039/D0TA05343K). It is also being used in ongoing research and in remote learning at the University of Bath.

The DL_MONTE code is a new Monte Carlo simulation code produced by STFC in the UK. To the best of my knowledge, there are no open source, research software codes that are aimed at analysis of DL_MONTE simulations. So all DL_MONTE functionality present in polypy is unique in the community.

MDAnalysis is the most widely used molecular dynamics analysis code and some of the functionality in polypy is already present in mdanalysis. The MDAnalysis.analysis.lineardensity module calculates the charge density in different dimensions, although according to the documentation, is limited to orthorombic, fixed volume cells. Polypy is designed to handle all cell types and will work for simulations ensembles other than NPT. Furthermore, the calculation of the electric field and electrostatic potential is unique to polypy.

MDAnalysis and polypy are both capable of calculating mean squared displacements. polypy goes a step further by allowing the calculation of diffusion coefficients and conductivities within localised regions of a structure e.g a grain boundary or surface. This method was used in two papers to calculate the diffusion coefficients the grain boundary and bulk regions of a simulation cell (10.1088/2515-7655/ab28b5, 10.1098/rsta.2019.0026).

DL_ANALYSER is also available for the analysis of DL_POLY simulations, however is only available under license
