### 2020-10-09 #2709 @kthyng
@symmy596 Your submission a bit borderline, but let's proceed. I encourage you to take the opportunity of your future reviewers suggestions to improve your software where possible so that it can be used by others in the future.
