### 2020-10-21 #2709 @richardjgowers
@symmy596 yeah you can edit the paper (& code) freely until it's published
