### 2020-11-08 #2824 @whedon
[REVIEW]: polypy - Analysis Tools for Solid State Molecular Dynamics and Monte Carlo Trajectories
**Submitting author:** @symmy596 (<a href="http://orcid.org/0000-0001-6059-497X">Adam Symington</a>)
**Repository:** <a href="https://github.com/symmy596/Polypy" target ="_blank">https://github.com/symmy596/Polypy</a>
**Version:** 0.8
