### 2020-11-27 #2824 @hmacdope
Hi @symmy596,
I'm one of your friendly JOSS reviewers.

Just a few things I thought I would point out as I go through, ranging in importance.

* The Latex in the docs did not build out of the box for me. Could be me but is worth investigating with a fresh conda environment.
* I couldn't see any documentation around how to run the tests although I may have missed it.
* The tests for the regional MSD seemed to be incomplete.
* **Important**: The files to run the tutorial notebooks out of the box are missing from `examples/example_data`. I'm guessing this is a problem with packaging (I cloned from master). I can't assess functionality until this is fixed.
* Suggestion: It would be great if there was an easy way to use only a section of the  MSD (the "middle"). Long story short, we want the middle portion of the MSD, after the initial ballistic motion is over and before the poor averaging regime takes over. See the discussion in [this](https://www.livecomsjournal.org/article/6324-best-practices-for-computing-transport-properties-1-self-diffusivity-and-viscosity-from-equilibrium-molecular-dynamics-article-v1-0) paper. This kind of behaviour is clear in your `fluorine diffusion in CaF_2` figure. This is not a blocking requirement.
* Nitpick: Would be fantastic if you could update the text describing the installation requirements in 'README.md' to include coveralls, coverage and sphinx etc as its in `requirements.txt`.

I will wait for you to respond here first, then raise the more pressing of these as issues on the central repo.

Onto the paper, it's looking good overall, just a few things.

* You are missing a state of the field section. There are a several packages aimed at analysing MD trajectories (MC less so). How do you differ from these?
* I think the second paragraph starting with `A molecular dynamics trajectory is a snapshot...` could be clearer and more accurate. A more detailed explanation around what MD and MC are is required for a general audience to follow what kind of data we are dealing with here.
* A description of the theory used to calculate some of the properties I think would be helpful to many. Currently it is just stated that you can compute them. I know it's in the docs, but it is central to the package.

Tracking well so far.
