### 2020-11-29 #2824 @symmy596
Hi @hmacdope
Firstly, thank you for reviewing the code, I really appreciate you taking the time to look through it.

When developing the code I had several large example trajectories stored in the examples folder and added that folder to the .gitignore to stop them all being written to github until I decided which ones I would use for the examples. I have rectified this and the files should now be in the repository. Apolgoies for the mishap.

With regards to the rest of the issues that you have raised, I will wait for you to complete your review and address all remaining issues in one go.

Once again, thank you for reviewing the code and I look forward to hearing how it can be improved.
