### 2020-12-03 #2824 @lscalfi
Hi @symmy596,
This is a useful package for doing analysis on MD or MC trajectories, the documentation is particularly detailed and I find the plotting functions especially aesthetic! Here are some points you may want to improve.

- I managed installing Polypy without problems although the required package section of the README should be updated with the list in `requirements.txt`. It could maybe be useful to specify if some packages are optional, and also to have the list of commands for those who use `conda` instead of `pip`.

- The documentation is well written and going in detail through different examples which makes it easy to get started with Polypy.
This is why I think it would be useful if the `Documentation` section of the README was a bit more detailed and highlighted the existence of such documentation (so that unexperienced users doesn't miss it).
By the way, @hmacdope it worked fine for me both with html and latexpdf.
However I ran into some trouble while going through the documentation. Overall I think there are discrepancies between the docs and the Jupyter notebooks (which seem to work).  So here is a non-ordered list of remarks about the documentation:

- I am not used to DLPOLY or DL_MONTE files such as HISTORY, CONFIG or ARCHIVE, but I would be interested to use these analysis tools all the same. Adding either a small description of the files or a link to a description could be very useful to adapt this to other trajectories format (such as .xyz, .pdb...).
- In the tutorials section, several times the files `../example_data/HISTORY` and `../example_data/ARCHIVE` are used but they don't exist, I guessed I should use the other versions but this should be corrected.
- The description of the `polypy.read.Trajectory` is lacking some elements: the `atom_labels` are instead `atom_list` and `atom_name`, and the is additional fields: `atoms_at_timestep`, `record_number`, `simulation_timestep`, `time`. Could the description of all the fields be provided? (I don't believe they are present in the API)
- The example on the Cerium oxide also contains several errors, I think there has been a mix of the CaF2 system and the CeO2 (you read `ca_density` but then use `ce_density`, same for `f_density` instead of `o_density`, and later the `fx_2d` is not defined).
**Important:** when trying to do the density profiles of `HISTORY_GB` which correspond to the CeO2, I had a Memory Error and could not do the analysis: does this happen often? Are you storing several large arrays ?
- The labels of the plots could be improved: the densities have no unit, the electric field should not be in V and I think the X or Y labels do not correspond to the x, y or z directions of the cell. I guess this is because of how the calculation is done but it would be best if the axis label corresponded to the required cell direction.
- Regarding the two-dimensional density plots, the plots are very nice. And the possibility of overlaying multiple plots is very useful for visualization purposes. Is the dimension you specify the normal vector to the plane? It would also be very useful to only have the density profile only in a slice (for example only for atoms between z=0 and z=5).
Would it be easy to do more complex geometries?. For example, is it possible to do an oblique cut (i.e. do a projection on a plane which is not the xy, xz or yx plane) to look at another crystal plane? It could also be nice to have projection on curved surfaces if analysing nanotubes (as done here for example https://pubs.acs.org/doi/10.1021/acs.langmuir.8b01115)
- Several functions have missing arguments (`charge_density` in the Two-dimensional... section, `combined_density_plot_multiple_species` in All together, and later `conductivity` in the ionic conductivity section) which trigger errors.
- Regarding the MSD: this functionality is also very useful. Can you define what are the `sweeps` in the documentation? By reading the API, the values are obtained by only the initial frame as reference by default, but it would probably be best to average over all starting timesteps. It is also usually useful to discard the initial ballistic regime and the last points where you have less statistics when computing the diffusion coefficient, is it possible in this case or the fit is done on the whole time range?
- For the regional MSD, it is not clear to me how to specify the region in space? Can it only be a slice or also only a cube by specifying xlow, ylow, zlow, xhi, yhi and zhi?

- Usually there is a first equilibration phase in MD or MC, is it easily possible to specify from which timestep on the trajectory should be analyzed? This could be an additional point to add to the tutorials, especially for the density calculations.
Another widely used functionality is the radial distribution function which should be easily computed using this package. It would be interesting to add it.

- I didn't see any documentation about how to run the tests in the /tests folder.

- About the paper itself, I noticed the acknowledgments are different from those in the README, should it be updated? Lastly, there are several other packages that do similar (sometimes more extensive) analysis, could you point out the differences and add a state-of-the-art section in the paper? For example, I used Chemfiles (chemfiles.org) and cfiles (https://github.com/chemfiles/cfiles) (written in C or C++).
