### 2020-12-03 #2824 @hmacdope
@symmy596 I will also add to @lscalfi's comment on the MSD. The MSD is normally defined as an ensemble average over all possible lag times. Do you average over all possible lag times and is this related to the `sweeps` parameter? Just for clarification. 👍
