+++ paper/paper.bib
@@ -141,3 +141,34 @@ year  = {2013},
 publisher = {Taylor & Francis},
 doi = {10.1080/08927022.2013.839871}
 }
+
+@software{chemfiles,
+  author       = {Guillaume Fraux and
+                  Jonathan Fine and
+                  ezavod and
+                  German P. Barletta and
+                  Laura Scalfi and
+                  Mykola Dimura},
+  title        = {chemfiles/chemfiles: Version 0.9.3},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {0.9.3},
+  doi          = {10.5281/zenodo.3653157},
+  url          = {https://doi.org/10.5281/zenodo.3653157}
+}
+
+@article{mdanalysis,
+author = {Michaud-Agrawal, Naveen and Denning, Elizabeth J. and Woolf, Thomas B. and Beckstein, Oliver},
+title = {MDAnalysis: A toolkit for the analysis of molecular dynamics simulations},
+journal = {Journal of Computational Chemistry},
+volume = {32},
+number = {10},
+pages = {2319-2327},
+keywords = {molecular dynamics simulations, analysis, proteins, object-oriented design, software, membrane systems, Python programming language},
+doi = {https://doi.org/10.1002/jcc.21787},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
+year = {2011}
+}
+
