### 2021-02-07 #2824 @richardjgowers
Hey @symmy596 just reading through the paper and saw a couple things in the statement of need section:
- the references to DL_ANALYSER and Chemfiles probably want to go up to the other references of existing packages
- I'm pretty sure most packages will handle non-orthogonal unit cells unless there's a particular geometry I've missed?
