+++ paper/paper.bib
@@ -172,3 +172,20 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
 year = {2011}
 }
 
+
+@Article{dlanalyser,
+AUTHOR = {Yong, Chin W. and Todorov, Ilian T.},
+TITLE = {DL_ANALYSER Notation for Atomic Interactions (DANAI): A Natural Annotation System for Molecular Interactions, Using Ethanoic Acid Liquid as a Test Case},
+JOURNAL = {Molecules},
+VOLUME = {23},
+YEAR = {2018},
+NUMBER = {1},
+ARTICLE-NUMBER = {36},
+URL = {https://www.mdpi.com/1420-3049/23/1/36},
+PubMedID = {29295538},
+ISSN = {1420-3049},
+DOI = {10.3390/molecules23010036}
+}
+
+
+
