+++ paper/paper.md
@@ -52,9 +52,9 @@ The repository contains examples of the core functionality as well as tutorials,
 
 # Statement of Need
 
-There are a number of codes designed to analyse molecular dynamics trajectories that currently exist. `MDAnalysis` [@mdanalysis] is the most widely used molecular dynamics analysis code and some of the functionality in `polypy` is already present in `MDAnalysis`. The `MDAnalysis.analysis.lineardensity` module calculates the charge density in different dimensions, although according to the documentation, is limited to orthorombic, fixed volume cells. `polypy` is designed to handle all cell types and will work for several simulation ensembles including, NPT, NVT, semi grand and grand canonical. Furthermore, the calculation of the electric field and electrostatic potential is unique to `polypy`. `MDAnalysis` and `polypy` are both capable of calculating mean squared displacements. `polypy` goes a step further by allowing the calculation of diffusion coefficients and conductivities within localised regions of a structure e.g a grain boundary [@UO2: @CeO2] or local structural environments [@LLTO]. `polypy` is also unique in the sense that it is designed for the analysis of both molecular dynamics and Monte Carlo trajectories.
+There are a number of codes designed to analyse molecular dynamics trajectories that currently exist [@mdanalysis; @chemfiles; @dlanalyser]. `MDAnalysis` [@mdanalysis] is the most widely used molecular dynamics analysis code and some of the functionality in `polypy` is already present in `MDAnalysis`. The `MDAnalysis.analysis.lineardensity` module calculates the charge density in different dimensions, although according to the documentation, is limited to orthorombic, fixed volume cells. `polypy` is designed to work with several simulation ensembles including, NPT, NVT, semi grand and grand canonical. Furthermore, the calculation of the electric field and electrostatic potential is unique to `polypy`. `MDAnalysis` and `polypy` are both capable of calculating mean squared displacements. `polypy` goes a step further by allowing the calculation of diffusion coefficients and conductivities within localised regions of a structure e.g a grain boundary [@UO2: @CeO2] or local structural environments [@LLTO]. `polypy` is also unique in the sense that it is designed for the analysis of both molecular dynamics and Monte Carlo trajectories.
 
-There are a number of other packages that are available for the analysis of MD trajectories, DL_ANALYSER is also available under license for the analysis of DL_POLY simulations and chemfilles is available for the analysis of a wide range of file types [@chemfiles].
+There are a number of other packages that are available for the analysis of MD trajectories, DL_ANALYSER [@dlanalyser] is also available under license for the analysis of DL_POLY simulations and chemfilles is available for the analysis of a wide range of file types [@chemfiles].
 
 In summary the features that are unique to `polypy` are as follows
 
