### 2021-02-07 #2824 @symmy596
@richardjgowers
- I have updated the references and moved them to the start of the paragraph.
- Fair point. I have dropped this sentence.
