+++ paper/paper.bib
@@ -188,4 +188,14 @@ DOI = {10.3390/molecules23010036}
 }
 
 
+@Article{LLTO_2,
+author ="Symington, Adam and Molinari, Marco and Dawson, James Alexander and Statham, Joel and Purton, John and Canepa, Pieremanuele and Parker, Stephen",
+title  ="Elucidating the Nature of Grain Boundary Resistance in Lithium Lanthanum Titanate.",
+journal  ="J. Mater. Chem. A",
+year  ="2021",
+pages  ="-",
+publisher  ="The Royal Society of Chemistry",
+doi  ="10.1039/D0TA11539H",
+url  ="http://dx.doi.org/10.1039/D0TA11539H",
+abstract  ="Solid electrolytes for all-solid-state batteries are generating remarkable research interest as a means to improve{,} their safety{,} stability and performance. Solid electrolytes are often polycrystalline and the effect that grain boundaries have on the material properties is often not fully characterised. Here{,} we present a comprehensive molecular dynamics study that quantifies the effect of grain boundaries on Li-ion transport in perovskite Li3xLa(2/3)-xTiO3 (0 < x < 0.16) (LLTO). Our results predict that grain boundaries hinder Li-ion conductivity by 1 to 2 orders of magnitude compared to the bulk. We attribute the poor con-ductivity of the grain boundaries to significant structural alterations at the grain boundaries. Our detailed analysis provides important insight into the influence of grain boundary structure on transport of Li-ions in solid electrolyte materials."}
 
