### 2021-02-14 #2824 @symmy596
@richardjgowers
I think I need more clarification on your comments.
Q. the references to DL_ANALYSER and Chemfiles probably want to go up to the other references of existing packages.
A. I thought this referred to the references for different packages under the statement of need ad hence "There are a number of codes designed to analyse molecular dynamics trajectories that currently exist. MDAnalysis [@mdanalysis]" was altered to include the relavant citations and now reads, "There are a number of codes designed to analyse molecular dynamics trajectories that currently exist (Fraux et al., 2020;Michaud-Agrawal et al., 2011;Yong & Todorov, 2018)." (line 68).

Q. I'm pretty sure most packages will handle non-orthogonal unit cells unless there's a particular geometry I've missed?
A. I thought this was referring to, "polypy is designed to handle all cell types and will work for several simulation ensembles including, NPT, NVT, semi grand and grand canonical.". This sentence was changed and the reference to different cell types was removed (line74).
