### 2021-02-21 #2824 @richardjgowers
@symmy596 re references, in the above pdf, lines 68 & 82 should get merged.  re orthogonal cells line 88
