### 2021-02-27 #2824 @richardjgowers
@symmy596 ok I think we're good to go.  Couple of things, the DOI for the mdanalysis reference seems broken, it looks like there's two https:// pieces.  And you'll also have to create a zenodo DOI :

At this point could you:
