+++ paper/paper.bib
@@ -166,7 +166,7 @@ volume = {32},
 number = {10},
 pages = {2319-2327},
 keywords = {molecular dynamics simulations, analysis, proteins, object-oriented design, software, membrane systems, Python programming language},
-doi = {https://doi.org/10.1002/jcc.21787},
+doi = {10.1002/jcc.21787},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/jcc.21787},
 year = {2011}
