### 2021-02-28 #2824 @symmy596
@richardjgowers Thanks. I have fixed the doi for MDAnalysis.
I have made a tagged release. Tag 1.0.0
Zenodo record has been created, doi - 10.5281/zenodo.4568493

Thanks
Adam
