### 2021-03-03 #2824 @danielskatz
I've found some minor changes to be merged into the paper: https://github.com/symmy596/Polypy/pull/5
