### 2021-03-03 #2824 @danielskatz
Also, please carefully check the references, specifically the titles, to make sure they appear correctly.  I think "Monte Carlo" should be capitalized, for example, and some of the chemical names might also need changes.
