### 2021-03-03 #2824 @danielskatz
Use `@whedon generate pdf` as a new comment to regenerate, then let me know when you think that the references are all correct and the PR has been merged.
