+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{CeO2,
 author = {Adam R. Symington and Marco Molinari and Joel Statham and Ji Wu and Stephen C. Parker},
-title = {The role of dopant segregation on the oxygen vacancy distribution and oxygen diffusion in {CeO}2 grain boundaries},
+title = {The role of dopant segregation on the oxygen vacancy distribution and oxygen diffusion in {C}e{O}$_{2}$} grain boundaries},
 journal = {Journal of Physics: Energy},
 year = {2019},
 volume = {1},
@@ -12,7 +12,7 @@ url = {https://doi.org/10.1088%2F2515-7655%2Fab28b5}
 
 @article{UO2,
 author = {Adam R. Symington and Marco Molinari and Nicholas. A. Brincat and Nicholas R. Williams and Stephen C. Parker},
-title = {Defect segregation facilitates oxygen transport at fluorite UO<sub>2</sub> grain boundaries},
+title = {Defect segregation facilitates oxygen transport at fluorite {U}{O}$_{2}$} grain boundaries},
 journal = {Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences},
 year = {2019},
 volume = {377},
@@ -117,22 +117,22 @@ doi = {10.1149/1.1507597}
 }
 
 @article{Smith,
-title = "DL_POLY_2.0: A general-purpose parallel molecular dynamics simulation package",
-journal = "Journal of Molecular Graphics",
-volume = "14",
-number = "3",
-pages = "136 - 141",
-year = "1996",
-issn = "0263-7855",
-doi = "10.1016/S0263-7855(96)00043-4",
-url = "http://www.sciencedirect.com/science/article/pii/S0263785596000434",
-author = "W. Smith and T.R. Forester",
-abstract = "DL_POLY_2.0 is a general-purpose parallel molecular dynamics simulation package developed at Daresbury Laboratory under the auspices of the Council for the Central Laboratory of the Research Councils. Written to support academic research, it has a wide range of applications and is designed to run on a wide range of computers: from single processor workstations to parallel supercomputers. Its structure, functionality, performance, and availability are described."
+title = {DL\_POLY 2.0: A general-purpose parallel molecular dynamics simulation package},
+journal = {Journal of Molecular Graphics},
+volume = {14},
+number = {3},
+pages = {136 - 141},
+year = {1996},
+issn = {0263-7855},
+doi = {10.1016/S0263-7855(96)00043-4},
+url = {http://www.sciencedirect.com/science/article/pii/S0263785596000434},
+author = {W. Smith and T.R. Forester},
+abstract = {DL\_POLY 2.0 is a general-purpose parallel molecular dynamics simulation package developed at Daresbury Laboratory under the auspices of the Council for the Central Laboratory of the Research Councils. Written to support academic research, it has a wide range of applications and is designed to run on a wide range of computers: from single processor workstations to parallel supercomputers. Its structure, functionality, performance, and availability are described.}
 }
 
 @article{Purton,
 author = {J.A. Purton and J.C. Crabtree and S.C. Parker},
-title = {DL\_MONTE: a general purpose program for parallel Monte Carlo simulation},
+title = {DL\_MONTE: a general purpose program for parallel {M}onte {C}arlo simulation},
 journal = {Molecular Simulation},
 volume = {39},
 number = {14-15},
@@ -175,7 +175,7 @@ year = {2011}
 
 @Article{dlanalyser,
 AUTHOR = {Yong, Chin W. and Todorov, Ilian T.},
[…]
