+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{CeO2,
 author = {Adam R. Symington and Marco Molinari and Joel Statham and Ji Wu and Stephen C. Parker},
-title = {The role of dopant segregation on the oxygen vacancy distribution and oxygen diffusion in {C}e{O}$_{2}$} grain boundaries},
+title = {The role of dopant segregation on the oxygen vacancy distribution and oxygen diffusion in {CeO}$_2$ grain boundaries},
 journal = {Journal of Physics: Energy},
 year = {2019},
 volume = {1},
@@ -12,7 +12,7 @@ url = {https://doi.org/10.1088%2F2515-7655%2Fab28b5}
 
 @article{UO2,
 author = {Adam R. Symington and Marco Molinari and Nicholas. A. Brincat and Nicholas R. Williams and Stephen C. Parker},
-title = {Defect segregation facilitates oxygen transport at fluorite {U}{O}$_{2}$} grain boundaries},
+title = {Defect segregation facilitates oxygen transport at fluorite {UO}$_2$ grain boundaries},
 journal = {Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences},
 year = {2019},
 volume = {377},
